requirements.txt
supabase
streamlit
python-dotenv
statsapi
pytest
pandas
numpy